Fuzzing cert

+++ TLS-Core/src/test/java/de/rub/nds/tlsattacker/core/protocol/parser/extension/ClientCertificateUrlExtensionParserTest.java
@@ -34,6 +34,6 @@ public class ClientCertificateUrlExtensionParserTest {
         message = parser.parse();
 
         assertArrayEquals(extensionType.getValue(), message.getExtensionType().getValue());
-        assertEquals(extensionLength, (int) message.getExtensionLength().getValue());
+        assertEquals(extensionLength, (long) message.getExtensionLength().getValue());
     }
 }
