+++ TLS-Core/src/test/java/de/rub/nds/tlsattacker/core/protocol/preparator/extension/ClientCertificateUrlExtensionPreparatorTest.java
@@ -37,6 +37,6 @@ public class ClientCertificateUrlExtensionPreparatorTest {
         preparator.prepare();
 
         assertArrayEquals(extensionType.getValue(), message.getExtensionType().getValue());
-        assertEquals(extensionLength, (int) message.getExtensionLength().getValue());
+        assertEquals(extensionLength, (long) message.getExtensionLength().getValue());
     }
 }
